Desktop: WYSIWYG: Allow pasting images in editor
